fix CI: black format and make index-cache layout tests CPU-safe
Move pure MLA layout helpers out of aiter_mla so Pre Checkin unit tests
run without AITER, and skip builder tests when AITER is unavailable.
